Moved calling #setLiftOnScrollTargetView to MultiSelectListFragment
