+++ collect_app/src/main/java/org/odk/collect/android/formlists/savedformlist/DeleteSavedFormFragment.kt
@@ -11,8 +11,6 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.map
-import androidx.recyclerview.widget.RecyclerView
-import com.google.android.material.appbar.AppBarLayout
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import org.odk.collect.analytics.Analytics
 import org.odk.collect.android.R
@@ -48,8 +46,6 @@ class DeleteSavedFormFragment(
             }
         )
     }
-    private var appBarLayout: AppBarLayout? = null
-    private lateinit var list: RecyclerView
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
@@ -65,8 +61,7 @@ class DeleteSavedFormFragment(
                     it.empty.setTitle(getString(string.empty_list_of_forms_to_delete_title))
                     it.empty.setSubtitle(getString(string.empty_list_of_saved_forms_to_delete_subtitle))
 
-                    list = it.list
-                    list.addItemDecoration(RecyclerViewUtils.verticalLineDivider(context))
+                    it.list.addItemDecoration(RecyclerViewUtils.verticalLineDivider(context))
                 }
             }
             .build()
@@ -93,7 +88,6 @@ class DeleteSavedFormFragment(
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        appBarLayout = requireActivity().findViewById(org.odk.collect.androidshared.R.id.appBarLayout)
         menuHost?.addMenuProvider(
             SavedFormListListMenuProvider(requireContext(), savedFormListViewModel),
             viewLifecycleOwner,
@@ -109,11 +103,6 @@ class DeleteSavedFormFragment(
                 it.message = getString(string.form_delete_message)
             }
         }
-    }
-
-    override fun onResume() {
-        super.onResume()
-        appBarLayout?.setLiftOnScrollTargetView(list)
     }
 
     private fun onDeleteSelected(selected: LongArray) {
